+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## `6.x`
 
+- Add arithmetic primitives to `Util\Binary`: `increment()`, `decrement()` and
+  `addIntegerOffset()`. Over/underflowing throws new exception.
 - Add support for PHP `8.5` in README, and GitHub Action CI workflows.
 - Access protected method via Closure in unit tests, instead of reflection, to
   fix deprecation warnings in PHP `8.5`.